% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-addPhenotypes.R
\docType{methods}
\name{addPhenotypes}
\alias{addPhenotypes}
\alias{addPhenotypes,MagicGen-method}
\title{Add phenotypes to 'MagicGen' object}
\usage{
addPhenotypes(x, phenotypes, id)

\S4method{addPhenotypes}{MagicGen}(x, phenotypes, id)
}
\arguments{
\item{x}{an object of class 'MagicGen'.}

\item{phenotypes}{a 'data.frame' object with phenotypes.}

\item{id}{the column name containing the MAGIC IDs in the 'phenotypes' table.}
}
\value{
a 'MagicGenPhen' object with phenotypes added to initial object.
}
\description{
This function adds phenotypes to an existing 'MagicGen' object.
The genotypes are kept only for those individuals with a genotype.
}
\examples{
# to be added...
}
