% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadArabMagic.R
\name{downloadArabMagic}
\alias{downloadArabMagic}
\title{Download Arabidopsis MAGIC genotypes}
\usage{
downloadArabMagic(save_dir, tidy = TRUE, example_data = FALSE)
}
\arguments{
\item{save_dir}{directory to save all files.}

\item{tidy}{whether or not to "tidy" the retrieved files.
This will call the `tidyArabMagic()` function. Default: TRUE}

\item{example_data}{logical. Whether or not to download an example phenotype 
data.}
}
\value{
files are downloaded to a directory of choice.
}
\description{
This function downloads SNP genotypes for 1254 SNPs typed with GoldenGate
technology and reported in Kover et. al (2009). Files are downloaded from
http://mtweb.cs.ucl.ac.uk/mus/www/magic/
}
\examples{
\dontrun{
downloadArabMagic("~/temp/magic_snp_1k")
}
}
