% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-estimateFounderEffect.R
\docType{methods}
\name{estimateFounderEffect}
\alias{estimateFounderEffect}
\alias{estimateFounderEffect,MagicGenPhen-method}
\title{Estimate founder effect based on probabilistic sampling}
\usage{
estimateFounderEffect(x, phenotype, marker, n_samples = 500)

\S4method{estimateFounderEffect}{MagicGenPhen}(x, phenotype, marker,
  n_samples = 500)
}
\arguments{
\item{x}{an object of class MagicData.}

\item{phenotype}{the phenotype to get the results from.}

\item{marker}{the SNP marker to estimate the effects for.}

\item{n_samples}{number of Monte Carlo samples to draw.}
}
\value{
a data.frame with phenotype estimates for each founder accession.
}
\description{
Estimate founder effect based on probabilistic sampling
}
\examples{
...
}
