% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\name{getFounderGenotypes}
\alias{getFounderGenotypes}
\title{Retrive SNP genotypes of MAGIC founders from a MagicGen object}
\usage{
getFounderGenotypes(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a list of genotypes for each founder accession.
}
\description{
Retrive SNP genotypes of MAGIC founders from a MagicGen object
}
