% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\docType{methods}
\name{getGenotypes}
\alias{getGenotypes}
\alias{getGenotypes,MagicGen-method}
\title{Retrieve genotypes from a MagicGen object}
\usage{
getGenotypes(x, type = c("probability", "allele"))

\S4method{getGenotypes}{MagicGen}(x, type = c("probability", "allele"))
}
\arguments{
\item{x}{an object of class MagicGen}

\item{type}{which type of genotype to retrieve.}
}
\value{
genotypes from MagicGenPhen object
}
\description{
This function can be used to extract the genotypes for all the MAGIC lines in
a MagicGen or MagicGenPhen object. The genotypes can either be the SNP genotypes
or the probability matrix of genotype ancestry for each MAGIC line.
}
