% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\name{getSnpGenotypes}
\alias{getSnpGenotypes}
\title{Retrive SNP genotypes of the MAGIC lines from a MagicGen object}
\usage{
getSnpGenotypes(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a list of genotypes SNP.
}
\description{
Retrive SNP genotypes of the MAGIC lines from a MagicGen object
}
