% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-inferMagicFounder.R
\docType{methods}
\name{inferMagicFounder}
\alias{inferMagicFounder}
\alias{inferMagicFounder,MagicGenPhen-method}
\title{Infer most likely founder for a given SNP}
\usage{
inferMagicFounder(x, marker, prob_thr = 0.5)

\S4method{inferMagicFounder}{MagicGenPhen}(x, marker, prob_thr = 0.5)
}
\arguments{
\item{x}{an object of class MagicData.}

\item{marker}{name of SNP marker to infer founder.}

\item{prob_thr}{probability threshold for attributing a genotype to a founder.}
}
\value{
a data.frame with the most likely founder for each MAGIC line.
}
\description{
Infer most likely founder for a given SNP
}
\examples{
# to be added...
}
