% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicFounderReconstruct.R
\name{magicFounderReconstruct}
\alias{magicFounderReconstruct}
\title{Calculate MAGIC founder genotype probabilities}
\usage{
magicFounderReconstruct(ped, map, alleles, phenotypes, id)
}
\arguments{
\item{ped}{genotype file (.happy.ped).}

\item{map}{map file (.happy.map).}

\item{alleles}{alleles file (.happy.alleles).}

\item{phenotypes}{a data.frame with phenotypes and with one of the columns being the MAGIC line IDs.}

\item{id}{the name of the variable with IDs in the data.frame passed to `phenotypes` argument.}
}
\value{
an object of type "MagicGen"
}
\description{
Calculate MAGIC founder genotype probabilities
}
\examples{
...
}

