% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-magicQtlScan.R
\docType{methods}
\name{magicQtlScan}
\alias{magicQtlScan}
\alias{magicQtlScan,MagicGenPhen-method}
\title{Make a QTL scan}
\usage{
magicQtlScan(x, phenotype, covariates = NULL, snp_cond = NULL, h1 = NULL,
  h0 = NULL, perm = 0, cores = 1)

\S4method{magicQtlScan}{MagicGenPhen}(x, phenotype, covariates = NULL,
  snp_cond = NULL, h1 = NULL, h0 = NULL, perm = 0, cores = 1)
}
\arguments{
\item{x}{an object of class MagicGenPhen}

\item{phenotype}{to run QTL analysis on.}

\item{covariates}{optional name of column(s) from phenotypes to use as covariate(s). Default: NULL}

\item{snp_cond}{optional ID of marker(s) to use as covariate(s) in the model. Default: NULL}

\item{h1}{optional optional model to test. Need to provide `h0` as well. Default: NULL}

\item{h0}{optional optional null model to test against. Need to provide `h1` as well. Default: NULL}

\item{perm}{optional number of permutations to run. If provided will permute the phenotypes 
the number of requested times and then run the QTL scan. It will add a new column to
the output with a genome-wide p-value (proportion of times where observed p > permuted p). Default: 0}

\item{cores}{number of cores to use (Windows only supports 1 core). Default: 1}
}
\value{
a data.frame with the QTL results.
}
\description{
Make a QTL scan
}
\examples{
...
}

