% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyArabMagic.R
\name{tidyArabMagic}
\alias{tidyArabMagic}
\title{Tidy Arabidopsis MAGIC genotypes}
\usage{
tidyArabMagic(snp_dir)
}
\arguments{
\item{snp_dir}{directory where SNP genotypes are stored. These files can be 
obtained with the `downloadMagicSnp()` function.}
}
\description{
Tidy Arabidopsis MAGIC genotypes
}
\examples{
downloadArabMagic('~/temp/magic_snp_1k/', tidy = FALSE)
tidyArabMagic('~/temp/magic_snp_1k/')
}

