% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.class-MagicGen.R
\docType{class}
\name{MagicGen-class}
\alias{MagicGen}
\alias{MagicGen-class}
\title{MagicGen class}
\description{
MagicGen class
}
\section{Slots}{

\describe{
\item{\code{phenotypes}}{data.frame with phenotypes for each MAGIC line.}

\item{\code{markers}}{data.frame with marker names and positions.}

\item{\code{genotypes}}{list with genotype probability matrices.}
}}

