% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\docType{methods}
\name{getFounderGenotypes}
\alias{getFounderGenotypes}
\alias{getFounderGenotypes,MagicGen-method}
\title{Retrieve SNP genotypes of MAGIC founders from a MagicGen object}
\usage{
getFounderGenotypes(x)

\S4method{getFounderGenotypes}{MagicGen}(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a list of genotypes for each founder accession.
}
\description{
Retrieve SNP genotypes of MAGIC founders from a MagicGen object
}
