% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\name{getGenotypes}
\alias{getGenotypes}
\title{Retrive genotypes from a MagicGen object}
\usage{
getGenotypes(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a list of genotype probabilities for every SNP.
}
\description{
Retrive genotypes from a MagicGen object
}

