% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\docType{methods}
\name{getMarkers}
\alias{getMarkers}
\alias{getMarkers,MagicGen-method}
\title{Retrieve markers from a MagicGen object}
\usage{
getMarkers(x)

\S4method{getMarkers}{MagicGen}(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a data.frame of marker names and positions.
}
\description{
Retrieve markers from a MagicGen object
}
