% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Retrive phenotypes from a MagicGen object}
\usage{
getPhenotypes(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
data.frame of phenotypes
}
\description{
Retrive phenotypes from a MagicGen object
}

