% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\docType{methods}
\name{getPhenotypes}
\alias{getPhenotypes}
\alias{getPhenotypes,MagicGenPhen-method}
\title{Retrieve phenotypes from a MagicGenPhen object}
\usage{
getPhenotypes(x)

\S4method{getPhenotypes}{MagicGenPhen}(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
data.frame of phenotypes
}
\description{
Retrieve phenotypes from a MagicGenPhen object
}
