% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class-MagicGen-accessors.R
\name{getProbGenotypes}
\alias{getProbGenotypes}
\title{Retrieve genotype probabilities for the MAGIC lines from a MagicGen object}
\usage{
getProbGenotypes(x)
}
\arguments{
\item{x}{an object of class "MagicGen".}
}
\value{
a list of genotype probabilities for every SNP.
}
\description{
Retrieve genotype probabilities for the MAGIC lines from a MagicGen object
}
