% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicFounderReconstruct.R
\name{magicFounderReconstruct}
\alias{magicFounderReconstruct}
\title{Calculate MAGIC founder genotype probabilities}
\usage{
magicFounderReconstruct(snp_dir, phenotypes = NULL, id = NULL)
}
\arguments{
\item{snp_dir}{directory with all genotype files. The function will be looking 
for files named "all_chr.MAGIC.happy.data", "all_chr.MAGIC.happy.map", 
"all_chr.MAGIC.happy.alleles", "founder_genotypes.tsv" and "magic_genotypes.tsv". 
All these files can be created with either the \code{tidyArabMagic} or
\code{downloadArabMagic} functions.}

\item{phenotypes}{a data.frame with phenotypes and with one of the columns being the MAGIC line IDs. 
This is set to NULL by default, which returns a 'MagicGen' object with no phenotypes included.}

\item{id}{the name of the variable with IDs in the data.frame passed to `phenotypes` argument.
This is set to NULL by default, which returns a 'MagicGen' object with no phenotypes included.}
}
\value{
an object of type "MagicGen" or "MagicGenPhen".
}
\description{
This function calls the function `happy.hbrem::happy()` to calculate 
the probability matrix of descent from each founder of the MAGIC lines
at each marker SNP.
}
\examples{
...
}
