% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-scanQtl.R
\name{magicQtlScan}
\alias{magicQtlScan}
\title{Make a QTL scan - deprecated function}
\usage{
magicQtlScan(...)
}
\arguments{
\item{...}{placeholder for any options given by user}
}
\description{
Make a QTL scan - deprecated function
}
